requirements.txt
beautifulsoup4
imagehash
instagrapi
instaloader
joblib
numpy
pandas
pillow
requests
scikit-learn
selenium
streamlit
webdriver-manager
